# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/plans": {
            "get": {
                "tags": [
                    "plan"
                ],
                "summary": "get plans / get plans for a specific month.",
                "description": "get plans / get plans for a specific month.",
                "operationId": "68b204e5906e1c5bd239842e172d8481",
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "description": "search by month(01,02,03,04....ex) example:2021-10",
                        "required": false
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "search by page",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "plan"
                ],
                "summary": "create new plan",
                "operationId": "create new plan",
                "requestBody": {
                    "description": "create new plan",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "date": {
                                        "type": "string",
                                        "example": "2021-12-01"
                                    },
                                    "exercises": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/plans/{plan_id}": {
            "get": {
                "tags": [
                    "plan"
                ],
                "summary": "show plan by id",
                "description": "show plan by id",
                "operationId": "7f1821346b5ec4d6c58ea80b6afed0fa",
                "parameters": [
                    {
                        "name": "plan_id",
                        "in": "path",
                        "description": "show plan by id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "plan"
                ],
                "summary": "edit plan",
                "description": "edit plan",
                "operationId": "3a2553e7294a5fa592f58d9f48c605ae",
                "parameters": [
                    {
                        "name": "plan_id",
                        "in": "path",
                        "description": "plan_id",
                        "required": true
                    }
                ],
                "requestBody": {
                    "description": "create new plan",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "plandate",
                                    "exercises"
                                ],
                                "properties": {
                                    "exercises": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "plan"
                ],
                "summary": "delete plan by id",
                "description": "delete plan by id",
                "operationId": "07ac350859fda8ac1877b83e11907962",
                "parameters": [
                    {
                        "name": "plan_id",
                        "in": "path",
                        "description": "delete plan by id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}